Enable more tests on universal poker and expose 2 bugs.
- Generate a walkthrough for the default parameters
- Expose 2 bugs when not using the default parameters:
  + The serialization segfaults (the parsing seems to be failing)
  + InformationStateTensorShape fails on an internal sanity check.

PiperOrigin-RevId: 283949077

+++ open_spiel/games/universal_poker/acpc_cpp/acpc_game.h
@@ -73,7 +73,7 @@ class ACPCGame {
   std::string ToString() const;
   bool IsLimitGame() const;
   uint8_t GetNbRounds() const;
-  uint8_t GetNbPlayers() const;
+  int GetNbPlayers() const;
   // Returns the number of private cards for each player in this game.
   uint8_t GetNbHoleCardsRequired() const;
   uint8_t GetNbBoardCardsRequired(uint8_t round) const;
